Extract custom matcher
